+++ a_matrix_local.cpp
@@ -3,12 +3,12 @@
 
 void a_matrix_local::fill_in_cell(cell* current, cell* above, cell* left, cell* above_left) {
 	//FIXME: is this right??
-	float d_score = d_matrix->cells[current->row][current->col].score;
-	float i_score = i_matrix->cells[current->row][current->col].score;
-	float substitution_score = substitution->get_score(sequence1.at(current->row - 1), sequence2.at(current->col - 1));
-	float a_score = above_left->score + substitution_score;
+	int d_score = d_matrix->cells[current->row][current->col].score;
+	int i_score = i_matrix->cells[current->row][current->col].score;
+	int substitution_score = substitution->get_score(sequence1.at(current->row - 1), sequence2.at(current->col - 1));
+	int a_score = above_left->score + substitution_score;
 
-	if(std::max(d_score, std::max(i_score, std::max(a_score, 0.0f))) == 0.0f) {
+	if(std::max(d_score, std::max(i_score, std::max(a_score, 0))) == 0) {
 		current->score = 0;
 		current->previous = 0;
 		return;
