+++ i_matrix.h
@@ -7,14 +7,14 @@ private:
 	matrix* a_matrix;
 
 public:
-	i_matrix(unsigned int size, substitution_matrix* substitution, float gap_open, float gap_extend) 
+	i_matrix(unsigned int size, substitution_matrix* substitution, int gap_open, int gap_extend) 
 		: matrix(size, substitution, gap_open, gap_extend) {};
 
 	void set_a_matrix(matrix* a_matrix);
 	friend class a_matrix;
 
 protected:
-	virtual float get_initial_score(int row, int col);
+	virtual int get_initial_score(int row, int col);
 	virtual cell* get_initial_pointer(int row, int col) { return 0; };
 	virtual void fill_in_cell(cell* current, cell* above, cell* left, cell* above_left);
